+++ source/Example.Orchestrations/Processes/BRS_X01/Example/V1/Activities/SecondStepStopActivity_Brs_X01_Example_V1.cs
@@ -34,10 +34,8 @@ internal class SecondStepStopActivity_Brs_X01_Example_V1(
             .GetAsync(input.OrchestrationInstanceId)
             .ConfigureAwait(false);
 
-        var step = orchestrationInstance.Steps.Single(x => x.Sequence == Orchestration_Brs_X01_Example_V1.SkippableStepSequence);
-
-        if (step.IsSkipped())
-            return;
+        var step = orchestrationInstance.Steps
+            .Single(step => step.Sequence == Orchestration_Brs_X01_Example_V1.SkippableStepSequence);
 
         step.Lifecycle.TransitionToTerminated(Clock, OrchestrationStepTerminationStates.Succeeded);
         await ProgressRepository.UnitOfWork.CommitAsync().ConfigureAwait(false);