+++ source/Example.Orchestrations/Processes/BRS_X01/Example/V1/Orchestration_Brs_X01_Example_V1.cs
@@ -14,6 +14,7 @@
 
 using Energinet.DataHub.Example.Orchestrations.Abstractions.Processes.BRS_X01.Example.V1.Model;
 using Energinet.DataHub.Example.Orchestrations.Processes.BRS_X01.Example.V1.Activities;
+using Energinet.DataHub.Example.Orchestrations.Processes.BRS_X01.Example.V1.Model;
 using Energinet.DataHub.ProcessManagement.Core.Domain.OrchestrationInstance;
 using Energinet.DataHub.ProcessManagement.Core.Infrastructure.Extensions.DurableTask;
 using Microsoft.Azure.Functions.Worker;
@@ -26,6 +27,13 @@ internal class Orchestration_Brs_X01_Example_V1
     internal const int FirstStepSequence = 1;
     internal const int SkippableStepSequence = 2;
 
+    private readonly TaskOptions _defaultRetryOptions;
+
+    public Orchestration_Brs_X01_Example_V1()
+    {
+        _defaultRetryOptions = CreateDefaultRetryOptions();
+    }
+
     [Function(nameof(Orchestration_Brs_X01_Example_V1))]
     public async Task<string> Run(
         [OrchestrationTrigger] TaskOrchestrationContext context)
@@ -36,45 +44,48 @@ internal class Orchestration_Brs_X01_Example_V1
             return "Error: No input specified.";
         }
 
-        var defaultRetryOptions = CreateDefaultRetryOptions();
+        var instanceId = new OrchestrationInstanceId(Guid.Parse(context.InstanceId));
 
         // Initialize
-        await context.CallActivityAsync(
+        var executionPlan = await context.CallActivityAsync<OrchestrationExecutionPlan>(
             nameof(InitializeOrchestrationActivity_Brs_X01_Example_V1),
             new InitializeOrchestrationActivity_Brs_X01_Example_V1.ActivityInput(
-                new OrchestrationInstanceId(Guid.Parse(context.InstanceId))),
-            defaultRetryOptions);
+                instanceId),
+            _defaultRetryOptions);
 
         // First Step
         await context.CallActivityAsync(
             nameof(FirstStepStartActivity_Brs_X01_Example_V1),
             new FirstStepStartActivity_Brs_X01_Example_V1.ActivityInput(
-                new OrchestrationInstanceId(Guid.Parse(context.InstanceId))),
-            defaultRetryOptions);
+                instanceId),
+            _defaultRetryOptions);
         await context.CallActivityAsync(
             nameof(FirstStepStopActivity_Brs_X01_Example_V1),
             new FirstStepStopActivity_Brs_X01_Example_V1.ActivityInput(
-                new OrchestrationInstanceId(Guid.Parse(context.InstanceId))),
-            defaultRetryOptions);
+                instanceId),
+            _defaultRetryOptions);
 
         // Skippable step
-        await context.CallActivityAsync(
-            nameof(SecondStepStartActivity_Brs_X01_Example_V1),
-            new SecondStepStartActivity_Brs_X01_Example_V1.ActivityInput(
-                new OrchestrationInstanceId(Guid.Parse(context.InstanceId))),
-            defaultRetryOptions);
-        await context.CallActivityAsync(
-            nameof(SecondStepStopActivity_Brs_X01_Example_V1),
-            new SecondStepStopActivity_Brs_X01_Example_V1.ActivityInput(
-                new OrchestrationInstanceId(Guid.Parse(context.InstanceId))),
-            defaultRetryOptions);
+        if (!executionPlan.SkippedStepsBySequence.Contains(SkippableStepSequence))
+        {
+            await context.CallActivityAsync(
+                nameof(SecondStepStartActivity_Brs_X01_Example_V1),
+                new SecondStepStartActivity_Brs_X01_Example_V1.ActivityInput(
+                    instanceId),
+                _defaultRetryOptions);
+            await context.CallActivityAsync(
+                nameof(SecondStepStopActivity_Brs_X01_Example_V1),
+                new SecondStepStopActivity_Brs_X01_Example_V1.ActivityInput(
+                    instanceId),
+                _defaultRetryOptions);
+        }
 
         // Terminate
         await context.CallActivityAsync(
             nameof(TerminateOrchestrationActivity_Brs_X01_Example_V1),
             new TerminateOrchestrationActivity_Brs_X01_Example_V1.ActivityInput(
-                new OrchestrationInstanceId(Guid.Parse(context.InstanceId))),
-            defaultRetryOptions);
+                instanceId),
+            _defaultRetryOptions);
 
         return "Success";
     }