+++ source/ProcessManager.Core/Domain/OrchestrationInstance/OrchestrationInstance.cs
@@ -83,6 +83,39 @@ public class OrchestrationInstance
     internal OrchestrationDescriptionId OrchestrationDescriptionId { get; }
 
     /// <summary>
+    /// Transition a step's lifecycle to running
+    /// </summary>
+    /// <param name="sequence">The sequence number of the step to transition</param>
+    /// <param name="clock"></param>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown if a step with the given <paramref name="sequence"/> isn't found.</exception>
+    public void TransitionStepToRunning(int sequence, IClock clock)
+    {
+        var step = Steps.SingleOrDefault(s => s.Sequence == sequence);
+
+        if (step == null)
+            throw new ArgumentOutOfRangeException(nameof(sequence), sequence, "A step with the given sequence does not exist");
+
+        step.Lifecycle.TransitionToRunning(clock);
+    }
+
+    /// <summary>
+    /// Transition a step's lifecycle to terminated, with the given <paramref name="terminationState"/>
+    /// </summary>
+    /// <param name="sequence">The sequence number of the step to transition</param>
+    /// <param name="terminationState">The state of the termination step (Succeeded, failed etc.)</param>
+    /// <param name="clock"></param>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown if a step with the given <paramref name="sequence"/> isn't found.</exception>
+    public void TransitionStepToTerminated(int sequence, OrchestrationStepTerminationStates terminationState, IClock clock)
+    {
+        var step = Steps.SingleOrDefault(s => s.Sequence == sequence);
+
+        if (step == null)
+            throw new ArgumentOutOfRangeException(nameof(sequence), sequence, "A step with the given sequence does not exist");
+
+        step.Lifecycle.TransitionToTerminated(clock, terminationState);
+    }
+
+    /// <summary>
     /// Factory method that ensures domain rules are obeyed when creating a new
     /// orchestration instance.
     /// </summary>